+++ Libplanet.Tests/Blockchain/BlockChainTest.cs
@@ -38,7 +38,7 @@ namespace Libplanet.Tests.Blockchain
         private StoreFixture _fx;
         private BlockPolicy<DumbAction> _policy;
         private BlockChain<DumbAction> _blockChain;
-        private ValidatingActionRenderer<DumbAction> _renderer;
+        private ValidatingActionRenderer _renderer;
         private Block _validNext;
         private IStagePolicy<DumbAction> _stagePolicy;
 
@@ -56,7 +56,7 @@ namespace Libplanet.Tests.Blockchain
                 getMaxTransactionsBytes: _ => 50 * 1024);
             _stagePolicy = new VolatileStagePolicy<DumbAction>();
             _fx = GetStoreFixture(_policy.BlockAction);
-            _renderer = new ValidatingActionRenderer<DumbAction>();
+            _renderer = new ValidatingActionRenderer();
             _blockChain = BlockChain<DumbAction>.Create(
                 _policy,
                 _stagePolicy,
@@ -70,9 +70,8 @@ namespace Libplanet.Tests.Blockchain
                     actionTypeLoader: new SingleActionLoader(typeof(DumbAction)),
                     feeCalculator: null
                 ),
-                renderers: new[] { new LoggedActionRenderer<DumbAction>(_renderer, Log.Logger) }
+                renderers: new[] { new LoggedActionRenderer(_renderer, Log.Logger) }
             );
-            _renderer.BlockChain = _blockChain;
             _renderer.ResetRecords();
 
             _validNext = _blockChain.EvaluateAndSign(
@@ -279,9 +278,9 @@ namespace Libplanet.Tests.Blockchain
             var store = new MemoryStore();
             var stateStore = new TrieStateStore(new MemoryKeyValueStore());
             var generatedRandomValueLogs = new List<int>();
-            IActionRenderer<DumbAction>[] renderers = Enumerable.Range(0, 2).Select(i =>
-                new LoggedActionRenderer<DumbAction>(
-                    new AnonymousActionRenderer<DumbAction>
+            IActionRenderer[] renderers = Enumerable.Range(0, 2).Select(i =>
+                new LoggedActionRenderer(
+                    new AnonymousActionRenderer
                     {
                         ActionRenderer = (act, context, nextStates) =>
                             // Consuming the random state through IRandom.Next() should not
@@ -316,8 +315,8 @@ namespace Libplanet.Tests.Blockchain
             var policy = new NullBlockPolicy<DumbAction>();
             var store = new MemoryStore();
             var stateStore = new TrieStateStore(new MemoryKeyValueStore());
-            var recordingRenderer = new RecordingActionRenderer<DumbAction>();
-            var renderer = new LoggedActionRenderer<DumbAction>(recordingRenderer, Log.Logger);
+            var recordingRenderer = new RecordingActionRenderer();
+            var renderer = new LoggedActionRenderer(recordingRenderer, Log.Logger);
             BlockChain<DumbAction> blockChain =
                 MakeBlockChain(policy, store, stateStore, renderers: new[] { renderer });
             var privateKey = new PrivateKey();
@@ -331,11 +330,9 @@ namespace Libplanet.Tests.Blockchain
             blockChain.Append(block, CreateBlockCommit(block));
 
             Assert.Equal(2, blockChain.Count);
-            IReadOnlyList<RenderRecord<DumbAction>.BlockEvent> blockLogs =
-                recordingRenderer.BlockRecords;
+            IReadOnlyList<RenderRecord.BlockEvent> blockLogs = recordingRenderer.BlockRecords;
             Assert.Equal(2, blockLogs.Count);
-            IReadOnlyList<RenderRecord<DumbAction>.ActionBase> actionLogs =
-                recordingRenderer.ActionRecords;
+            IReadOnlyList<RenderRecord.ActionBase> actionLogs = recordingRenderer.ActionRecords;
             Assert.Single(actions);
             Assert.Equal(prevBlock, blockLogs[0].OldTip);
             Assert.Equal(block, blockLogs[0].NewTip);
@@ -354,7 +351,7 @@ namespace Libplanet.Tests.Blockchain
             var store = new MemoryStore();
             var stateStore = new TrieStateStore(new MemoryKeyValueStore());
 
-            IActionRenderer<DumbAction> renderer = new AnonymousActionRenderer<DumbAction>
+            IActionRenderer renderer = new AnonymousActionRenderer
             {
                 ActionRenderer = (a, __, nextStates) =>
                 {
@@ -366,7 +363,7 @@ namespace Libplanet.Tests.Blockchain
                     }
                 },
             };
-            renderer = new LoggedActionRenderer<DumbAction>(renderer, Log.Logger);
+            renderer = new LoggedActionRenderer(renderer, Log.Logger);
             BlockChain<DumbAction> blockChain =
                 MakeBlockChain(policy, store, stateStore, renderers: new[] { renderer });
             var privateKey = new PrivateKey();
@@ -586,7 +583,7 @@ namespace Libplanet.Tests.Blockchain
                     blockAction: _policy.BlockAction);
 
                 store.PutBlock(genesis);
-                var renderer = new RecordingActionRenderer<DumbAction>();
+                var renderer = new RecordingActionRenderer();
                 var blockChain = BlockChain<DumbAction>.Create(
                     _policy,
                     new VolatileStagePolicy<DumbAction>(),
@@ -843,11 +840,11 @@ namespace Libplanet.Tests.Blockchain
             Assert.Empty(_blockChain.Store.IterateIndexes(previousChainId));
             Assert.Empty(_blockChain.Store.ListTxNonces(previousChainId));
 
-            RenderRecord<DumbAction>.BlockBase[] blockLevelRenders = _renderer.Records
-                .OfType<RenderRecord<DumbAction>.BlockBase>()
+            RenderRecord.BlockBase[] blockLevelRenders = _renderer.Records
+                .OfType<RenderRecord.BlockBase>()
                 .ToArray();
 
-            RenderRecord<DumbAction>.ActionBase[] actionRenders = _renderer.ActionRecords
+            RenderRecord.ActionBase[] actionRenders = _renderer.ActionRecords
                 .Where(r => IsDumbAction(r.Action))
                 .ToArray();
             DumbAction[] actions = actionRenders.Select(r => ToDumbAction(r.Action)).ToArray();
@@ -862,9 +859,9 @@ namespace Libplanet.Tests.Blockchain
             if (render)
             {
                 Assert.Equal(2, blockLevelRenders.Length);
-                Assert.IsType<RenderRecord<DumbAction>.BlockEvent>(blockLevelRenders[0]);
+                Assert.IsType<RenderRecord.BlockEvent>(blockLevelRenders[0]);
                 Assert.True(blockLevelRenders[0].Begin);
-                Assert.IsType<RenderRecord<DumbAction>.BlockEvent>(blockLevelRenders[1]);
+                Assert.IsType<RenderRecord.BlockEvent>(blockLevelRenders[1]);
                 Assert.True(blockLevelRenders[1].End);
 
                 Assert.True(blockLevelRenders[0].Index < actionRenders[0].Index);
@@ -879,7 +876,7 @@ namespace Libplanet.Tests.Blockchain
                 Assert.Equal("fork-bar", actions[1].Item);
                 Assert.Equal("fork-baz", actions[2].Item);
 
-                RenderRecord<DumbAction>.ActionBase[] blockActionRenders = _renderer.ActionRecords
+                RenderRecord.ActionBase[] blockActionRenders = _renderer.ActionRecords
                     .Where(r => IsMinerReward(r.Action))
                     .ToArray();
 
@@ -972,7 +969,7 @@ namespace Libplanet.Tests.Blockchain
         public void SwapForSameTip(bool render)
         {
             BlockChain<DumbAction> fork = _blockChain.Fork(_blockChain.Tip.Hash);
-            IReadOnlyList<RenderRecord<DumbAction>> prevRecords = _renderer.Records;
+            IReadOnlyList<RenderRecord> prevRecords = _renderer.Records;
             _blockChain.Swap(fork, render: render)();
 
             // Render methods should be invoked if and only if the tip changes
@@ -1481,7 +1478,6 @@ namespace Libplanet.Tests.Blockchain
         [SkippableFact]
         public void MakeTransactionWithSystemAction()
         {
-            var foo = Currency.Uncapped("FOO", 2, minters: null);
             var privateKey = new PrivateKey();
             Address address = privateKey.ToAddress();
             var action = new Initialize(
@@ -1636,8 +1632,7 @@ namespace Libplanet.Tests.Blockchain
             MakeIncompleteBlockStates(
                 IStore store,
                 IStateStore stateStore,
-                IRenderer<DumbAction> renderer = null
-            )
+                IRenderer renderer = null)
         {
             List<int> presentIndices = new List<int>() { 4, 7 };
             List<Block> presentBlocks = new List<Block>();
@@ -1846,9 +1841,9 @@ namespace Libplanet.Tests.Blockchain
             Assert.Empty(_renderer.BlockRecords);
             Block block = _blockChain.ProposeBlock(new PrivateKey());
             _blockChain.Append(block, CreateBlockCommit(block));
-            IReadOnlyList<RenderRecord<DumbAction>.BlockEvent> records = _renderer.BlockRecords;
+            IReadOnlyList<RenderRecord.BlockEvent> records = _renderer.BlockRecords;
             Assert.Equal(2, records.Count);
-            foreach (RenderRecord<DumbAction>.BlockEvent record in records)
+            foreach (RenderRecord.BlockEvent record in records)
             {
                 Assert.Equal(genesis, record.OldTip);
                 Assert.Equal(block, record.NewTip);