+++ Libplanet.Tests/Blockchain/Renderers/AtomicActionRendererTest.cs
@@ -12,15 +12,15 @@ namespace Libplanet.Tests.Blockchain.Renderers
 {
     public class AtomicActionRendererTest
     {
-        private readonly RecordingActionRenderer<Arithmetic> _record;
-        private readonly AtomicActionRenderer<Arithmetic> _renderer;
+        private readonly RecordingActionRenderer _record;
+        private readonly AtomicActionRenderer _renderer;
         private readonly IntegerSet _fx;
         private Transaction _successTx;
 
         public AtomicActionRendererTest()
         {
-            _record = new RecordingActionRenderer<Arithmetic>();
-            _renderer = new AtomicActionRenderer<Arithmetic>(_record);
+            _record = new RecordingActionRenderer();
+            _renderer = new AtomicActionRenderer(_record);
             _fx = new IntegerSet(new BigInteger?[] { 0 }, null, new[] { _renderer });
             (_successTx, _) = _fx.Sign(
                 0,
@@ -43,26 +43,26 @@ namespace Libplanet.Tests.Blockchain.Renderers
         public void Block()
         {
             _fx.Append(_fx.Propose());
-            IReadOnlyList<RenderRecord<Arithmetic>> records = _record.Records;
+            IReadOnlyList<RenderRecord> records = _record.Records;
             Assert.Equal(5, records.Count);
-            AssertTypeAnd<RenderRecord<Arithmetic>.BlockEvent>(
+            AssertTypeAnd<RenderRecord.BlockEvent>(
                 records[0], r => Assert.True(r.Begin));
-            AssertTypeAnd<RenderRecord<Arithmetic>.ActionSuccess>(records[1], r =>
+            AssertTypeAnd<RenderRecord.ActionSuccess>(records[1], r =>
             {
                 Assert.Equal(_successTx.Id, r.Context.TxId);
                 Assert.Equal(_successTx.Actions[0], r.Action);
             });
-            AssertTypeAnd<RenderRecord<Arithmetic>.ActionSuccess>(records[2], r =>
+            AssertTypeAnd<RenderRecord.ActionSuccess>(records[2], r =>
             {
                 Assert.Equal(_successTx.Id, r.Context.TxId);
                 Assert.Equal(_successTx.Actions[1], r.Action);
             });
-            AssertTypeAnd<RenderRecord<Arithmetic>.ActionSuccess>(records[3], r =>
+            AssertTypeAnd<RenderRecord.ActionSuccess>(records[3], r =>
             {
                 Assert.Equal(_successTx.Id, r.Context.TxId);
                 Assert.Equal(_successTx.Actions[2], r.Action);
             });
-            AssertTypeAnd<RenderRecord<Arithmetic>.BlockEvent>(records[4], r => Assert.True(r.End));
+            AssertTypeAnd<RenderRecord.BlockEvent>(records[4], r => Assert.True(r.End));
         }
 
         [Fact]
@@ -72,11 +72,11 @@ namespace Libplanet.Tests.Blockchain.Renderers
             _fx.Append(_fx.Propose());
             _record.ResetRecords();
             _fx.Chain.Swap(@base, true)();
-            IReadOnlyList<RenderRecord<Arithmetic>> records = _record.Records;
+            IReadOnlyList<RenderRecord> records = _record.Records;
             Assert.Equal(2, records.Count);
-            AssertTypeAnd<RenderRecord<Arithmetic>.BlockEvent>(
+            AssertTypeAnd<RenderRecord.BlockEvent>(
                 records[0], r => Assert.True(r.Begin));
-            AssertTypeAnd<RenderRecord<Arithmetic>.BlockEvent>(
+            AssertTypeAnd<RenderRecord.BlockEvent>(
                 records[1], r => Assert.True(r.End));
         }