+++ Libplanet.Tests/Blockchain/Renderers/LoggedActionRendererTest.cs
@@ -86,7 +86,7 @@ namespace Libplanet.Tests.Blockchain.Renderers
                     rehearsal
                 );
             Exception actionError = new Exception();
-            IActionRenderer<DumbAction> actionRenderer;
+            IActionRenderer actionRenderer;
             if (error)
             {
                 Action<IValue, IActionContext, Exception> render = (action, cxt, e) =>
@@ -103,7 +103,7 @@ namespace Libplanet.Tests.Blockchain.Renderers
                         throw new ThrowException.SomeException(string.Empty);
                     }
                 };
-                actionRenderer = new AnonymousActionRenderer<DumbAction>
+                actionRenderer = new AnonymousActionRenderer
                 {
                     ActionErrorRenderer = render,
                 };
@@ -124,13 +124,13 @@ namespace Libplanet.Tests.Blockchain.Renderers
                         throw new ThrowException.SomeException(string.Empty);
                     }
                 };
-                actionRenderer = new AnonymousActionRenderer<DumbAction>
+                actionRenderer = new AnonymousActionRenderer
                 {
                     ActionRenderer = render,
                 };
             }
 
-            actionRenderer = new LoggedActionRenderer<DumbAction>(
+            actionRenderer = new LoggedActionRenderer(
                 actionRenderer,
                 _logger,
                 LogEventLevel.Information
@@ -189,7 +189,7 @@ namespace Libplanet.Tests.Blockchain.Renderers
                 firstLog.Properties["BlockIndex"].ToString()
             );
             Assert.Equal(
-                $"\"{typeof(AnonymousActionRenderer<DumbAction>).FullName}\"",
+                $"\"{typeof(AnonymousActionRenderer).FullName}\"",
                 firstLog.Properties[Constants.SourceContextPropertyName].ToString()
             );
             Assert.Null(firstLog.Exception);
@@ -276,12 +276,12 @@ namespace Libplanet.Tests.Blockchain.Renderers
                 }
             }
 
-            IActionRenderer<DumbAction> actionRenderer = new AnonymousActionRenderer<DumbAction>
+            IActionRenderer actionRenderer = new AnonymousActionRenderer
             {
                 BlockRenderer = end ? (Action<DumbBlock, DumbBlock>)null : Callback,
                 BlockEndRenderer = end ? Callback : (Action<DumbBlock, DumbBlock>)null,
             };
-            actionRenderer = new LoggedActionRenderer<DumbAction>(actionRenderer, _logger);
+            actionRenderer = new LoggedActionRenderer(actionRenderer, _logger);
             var invoke = end
                 ? (Action<DumbBlock, DumbBlock>)actionRenderer.RenderBlockEnd
                 : actionRenderer.RenderBlock;
@@ -330,7 +330,7 @@ namespace Libplanet.Tests.Blockchain.Renderers
             );
             Assert.Equal($"\"{_genesis.Hash}\"", firstLog.Properties["OldHash"].ToString());
             Assert.Equal(
-                $"\"{typeof(AnonymousActionRenderer<DumbAction>).FullName}\"",
+                $"\"{typeof(AnonymousActionRenderer).FullName}\"",
                 firstLog.Properties[Constants.SourceContextPropertyName].ToString()
             );
             Assert.Null(firstLog.Exception);