+++ Libplanet.Tests/Blockchain/Renderers/LoggedRendererTest.cs
@@ -53,7 +53,7 @@ namespace Libplanet.Tests.Blockchain.Renderers
             bool called = false;
             LogEvent firstLog = null;
 
-            IRenderer<DumbAction> renderer = new AnonymousRenderer<DumbAction>
+            IRenderer renderer = new AnonymousRenderer
             {
                 BlockRenderer = (oldTip, newTip) =>
                 {
@@ -69,7 +69,7 @@ namespace Libplanet.Tests.Blockchain.Renderers
                     }
                 },
             };
-            renderer = new LoggedRenderer<DumbAction>(renderer, _logger);
+            renderer = new LoggedRenderer(renderer, _logger);
 
             Assert.False(called);
             Assert.Empty(LogEvents);
@@ -106,7 +106,7 @@ namespace Libplanet.Tests.Blockchain.Renderers
             );
             Assert.Equal($"\"{_genesis.Hash}\"", firstLog.Properties["OldHash"].ToString());
             Assert.Equal(
-                $"\"{typeof(AnonymousRenderer<DumbAction>).FullName}\"",
+                $"\"{typeof(AnonymousRenderer).FullName}\"",
                 firstLog.Properties[Constants.SourceContextPropertyName].ToString()
             );
             Assert.Null(firstLog.Exception);