+++ Libplanet/Blockchain/Renderers/AnonymousActionRenderer.cs
@@ -6,16 +6,16 @@ using Libplanet.Blocks;
 namespace Libplanet.Blockchain.Renderers
 {
     /// <summary>
-    /// An <see cref="IActionRenderer{T}"/> that invokes its callbacks.
-    /// In other words, this is an <see cref="IActionRenderer{T}"/> version of
-    /// <see cref="AnonymousRenderer{T}"/>.
+    /// An <see cref="IActionRenderer"/> that invokes its callbacks.
+    /// In other words, this is an <see cref="IActionRenderer"/> version of
+    /// <see cref="AnonymousRenderer"/>.
     /// <para>This class is useful when you want an one-use ad-hoc implementation (i.e., Java-style
-    /// anonymous class) of <see cref="IActionRenderer{T}"/> interface.</para>
+    /// anonymous class) of <see cref="IActionRenderer"/> interface.</para>
     /// </summary>
     /// <example>
     /// With object initializers, you can easily make an one-use action renderer:
     /// <code><![CDATA[
-    /// var actionRenderer = new AnonymousActionRenderer<ExampleAction>
+    /// var actionRenderer = new AnonymousActionRenderer
     /// {
     ///     ActionRenderer = (action, context, nextStates) =>
     ///     {
@@ -24,10 +24,7 @@ namespace Libplanet.Blockchain.Renderers
     /// };
     /// ]]></code>
     /// </example>
-    /// <typeparam name="T">An <see cref="IAction"/> type.  It should match to
-    /// <see cref="BlockChain{T}"/>'s type parameter.</typeparam>
-    public sealed class AnonymousActionRenderer<T> : AnonymousRenderer<T>, IActionRenderer<T>
-        where T : IAction, new()
+    public sealed class AnonymousActionRenderer : AnonymousRenderer, IActionRenderer
     {
         /// <summary>
         /// A callback function to be invoked together with
@@ -48,7 +45,7 @@ namespace Libplanet.Blockchain.Renderers
         public Action<Block, Block>? BlockEndRenderer { get; set; }
 
         /// <inheritdoc
-        /// cref="IActionRenderer{T}.RenderAction(IValue, IActionContext, IAccountStateDelta)"/>
+        /// cref="IActionRenderer.RenderAction(IValue, IActionContext, IAccountStateDelta)"/>
         public void RenderAction(
             IValue action,
             IActionContext context,
@@ -57,11 +54,11 @@ namespace Libplanet.Blockchain.Renderers
             ActionRenderer?.Invoke(action, context, nextStates);
 
         /// <inheritdoc
-        /// cref="IActionRenderer{T}.RenderActionError(IValue, IActionContext, Exception)"/>
+        /// cref="IActionRenderer.RenderActionError(IValue, IActionContext, Exception)"/>
         public void RenderActionError(IValue action, IActionContext context, Exception exception)
             => ActionErrorRenderer?.Invoke(action, context, exception);
 
-        /// <inheritdoc cref="IActionRenderer{T}.RenderBlockEnd(Block, Block)"/>
+        /// <inheritdoc cref="IActionRenderer.RenderBlockEnd(Block, Block)"/>
         public void RenderBlockEnd(Block oldTip, Block newTip) =>
             BlockEndRenderer?.Invoke(oldTip, newTip);
     }