+++ Libplanet/Blockchain/Renderers/Debug/RecordingActionRenderer.cs
@@ -12,39 +12,36 @@ namespace Libplanet.Blockchain.Renderers.Debug
     /// <summary>
     /// Records every render events.
     /// </summary>
-    /// <typeparam name="T">An <see cref="IAction"/> type.  It should match to
-    /// <see cref="Libplanet.Blockchain.BlockChain{T}"/>'s type parameter.</typeparam>
-    public class RecordingActionRenderer<T> : IActionRenderer<T>
-        where T : IAction, new()
+    public class RecordingActionRenderer : IActionRenderer
     {
-        private readonly List<RenderRecord<T>> _records;
+        private readonly List<RenderRecord> _records;
         private long _nextIndex;
 
         /// <summary>
-        /// Creates a new <see cref="RecordingActionRenderer{T}"/> instance.
+        /// Creates a new <see cref="RecordingActionRenderer"/> instance.
         /// </summary>
         public RecordingActionRenderer()
         {
             _nextIndex = 0;
-            _records = new List<RenderRecord<T>>();
+            _records = new List<RenderRecord>();
         }
 
         /// <summary>
         /// The list of recorded render events.
         /// </summary>
-        public IReadOnlyList<RenderRecord<T>> Records => _records;
+        public IReadOnlyList<RenderRecord> Records => _records;
 
-        internal IReadOnlyList<RenderRecord<T>.ActionBase> ActionRecords =>
-            Records.OfType<RenderRecord<T>.ActionBase>().ToImmutableArray();
+        internal IReadOnlyList<RenderRecord.ActionBase> ActionRecords =>
+            Records.OfType<RenderRecord.ActionBase>().ToImmutableArray();
 
-        internal IReadOnlyList<RenderRecord<T>.ActionSuccess> ActionSuccessRecords =>
-            Records.OfType<RenderRecord<T>.ActionSuccess>().ToImmutableArray();
+        internal IReadOnlyList<RenderRecord.ActionSuccess> ActionSuccessRecords =>
+            Records.OfType<RenderRecord.ActionSuccess>().ToImmutableArray();
 
-        internal IReadOnlyList<RenderRecord<T>.ActionError> ActionErrorRecords =>
-            Records.OfType<RenderRecord<T>.ActionError>().ToImmutableArray();
+        internal IReadOnlyList<RenderRecord.ActionError> ActionErrorRecords =>
+            Records.OfType<RenderRecord.ActionError>().ToImmutableArray();
 
-        internal IReadOnlyList<RenderRecord<T>.BlockEvent> BlockRecords =>
-            Records.OfType<RenderRecord<T>.BlockEvent>().ToImmutableArray();
+        internal IReadOnlyList<RenderRecord.BlockEvent> BlockRecords =>
+            Records.OfType<RenderRecord.BlockEvent>().ToImmutableArray();
 
         internal EventHandler<IValue>? RenderEventHandler { get; set; }
 
@@ -55,10 +52,10 @@ namespace Libplanet.Blockchain.Renderers.Debug
         {
             _nextIndex = 0;
             _records.Clear();
-            Log.Logger.ForContext<RecordingActionRenderer<T>>().Debug("Reset records");
+            Log.Logger.ForContext<RecordingActionRenderer>().Debug("Reset records");
         }
 
-        /// <inheritdoc cref="IActionRenderer{T}.RenderAction"/>
+        /// <inheritdoc cref="IActionRenderer.RenderAction"/>
         public virtual void RenderAction(
             IValue action,
             IActionContext context,
@@ -66,7 +63,7 @@ namespace Libplanet.Blockchain.Renderers.Debug
         )
         {
             _records.Add(
-                new RenderRecord<T>.ActionSuccess(
+                new RenderRecord.ActionSuccess(
                     index: _nextIndex++,
                     stackTrace: RemoveFirstLine(Environment.StackTrace).TrimEnd(),
                     action: action,
@@ -78,14 +75,14 @@ namespace Libplanet.Blockchain.Renderers.Debug
             RenderEventHandler?.Invoke(action, action);
         }
 
-        /// <inheritdoc cref="IActionRenderer{T}.RenderActionError"/>
+        /// <inheritdoc cref="IActionRenderer.RenderActionError"/>
         public virtual void RenderActionError(
             IValue action,
             IActionContext context,
             Exception exception
         ) =>
             _records.Add(
-                new RenderRecord<T>.ActionError(
+                new RenderRecord.ActionError(
                     index: _nextIndex++,
                     stackTrace: RemoveFirstLine(Environment.StackTrace).TrimEnd(),
                     action: action,
@@ -94,10 +91,10 @@ namespace Libplanet.Blockchain.Renderers.Debug
                 )
             );
 
-        /// <inheritdoc cref="IRenderer{T}.RenderBlock(Block, Block)"/>
+        /// <inheritdoc cref="IRenderer.RenderBlock(Block, Block)"/>
         public virtual void RenderBlock(Block oldTip, Block newTip) =>
             _records.Add(
-                new RenderRecord<T>.BlockEvent(
+                new RenderRecord.BlockEvent(
                     index: _nextIndex++,
                     stackTrace: RemoveFirstLine(Environment.StackTrace).TrimEnd(),
                     oldTip: oldTip,
@@ -105,10 +102,10 @@ namespace Libplanet.Blockchain.Renderers.Debug
                 )
             );
 
-        /// <inheritdoc cref="IActionRenderer{T}.RenderBlockEnd(Block, Block)"/>
+        /// <inheritdoc cref="IActionRenderer.RenderBlockEnd(Block, Block)"/>
         public virtual void RenderBlockEnd(Block oldTip, Block newTip) =>
             _records.Add(
-                new RenderRecord<T>.BlockEvent(
+                new RenderRecord.BlockEvent(
                     index: _nextIndex++,
                     stackTrace: RemoveFirstLine(Environment.StackTrace).TrimEnd(),
                     end: true,